misc(client): reoarganise SFML files & rename methods
