+++ client/external/sfml.hpp
@@ -25,8 +25,8 @@ namespace rtype
         public:
             virtual ~IGraphical() = default;
 
-            virtual void loadTextures() = 0;
-            virtual void draw(std::string sprite, int x, int y, float scale, int rotation, int size_x, int size_y) = 0;
+            virtual void loadAssets() = 0;
+            virtual void drawSprite(std::string sprite, int x, int y, float scale, int rotation, int size_x, int size_y) = 0;
             virtual void drawText(std::string text, int x, int y, int size) = 0;
             virtual void clear() = 0;
             virtual void display() = 0;
@@ -38,14 +38,14 @@ namespace rtype
             SFML();
             ~SFML();
 
-            KeyState handleEvents() override;
-            void loadTextures() override;
-            void draw(std::string sprite, int x, int y, float scale, int rotation, int size_x, int size_y) override;
+            void loadAssets() override;
+            void drawSprite(std::string sprite, int x, int y, float scale, int rotation, int size_x, int size_y) override;
             void drawText(std::string text, int x, int y, int size) override;
             void clear() override;
             void display() override;
-            void initTextures(std::string name, unsigned char const *sheet, unsigned int size);
+            KeyState handleEvents() override;
 
+            void initTextures(std::string name, unsigned char const *sheet, unsigned int size);
             sf::RenderWindow& getWindow() { return _window; };
 
         private:
